Update TypeScript & make scripts more DRY
