+++ package.json
@@ -4,19 +4,19 @@
   "version": "0.0.0",
   "scripts": {
     "bootstrap": "yarn install && lerna link",
-    "wsrun": "wsrun --bin $INIT_CWD/scripts/wsrun/bin.ts -p $PKG",
-    "dev": "$INIT_CWD/scripts/clean.ts && $INIT_CWD/scripts/dev.ts",
-    "build": "$INIT_CWD/scripts/clean.ts && $INIT_CWD/scripts/build.ts",
-    "clean": "$INIT_CWD/scripts/clean.ts",
-    "lint": "$INIT_CWD/scripts/lint.ts",
-    "test": "$INIT_CWD/scripts/test.ts",
-    "paths": "yarn --silent wsrun --parallel --no-prefix -r -c $INIT_CWD/scripts/wsrun/resolve-paths.ts",
+    "wsrun": "wsrun --bin $INIT_CWD/scripts/bin/wsrun/bin.ts -p $PKG",
+    "wsrun:paths": "yarn --silent wsrun --parallel --no-prefix -r -c $INIT_CWD/scripts/bin/wsrun/resolve-paths.ts",
+    "dev": "$INIT_CWD/scripts/bin/clean.ts && $INIT_CWD/scripts/bin/dev.ts",
+    "build": "$INIT_CWD/scripts/bin/clean.ts && $INIT_CWD/scripts/bin/build.ts",
+    "clean": "$INIT_CWD/scripts/bin/clean.ts",
+    "lint": "$INIT_CWD/scripts/bin/lint.ts",
+    "test": "$INIT_CWD/scripts/bin/test.ts",
     "postinstall": "husky install"
   },
   "devDependencies": {
     "@ikscodes/browser-env": "~0.3.1",
-    "@ikscodes/eslint-config": "~7.0.0",
-    "@ikscodes/prettier-config": "~2.0.0",
+    "@ikscodes/eslint-config": "~7.0.2",
+    "@ikscodes/prettier-config": "~2.0.1",
     "@istanbuljs/nyc-config-typescript": "~0.1.3",
     "@types/jest": "^27.0.0",
     "@types/jsdom": "~12.2.4",
@@ -47,7 +47,7 @@
     "npm-run-all": "^4.1.5",
     "nyc": "13.1.0",
     "ora": "~5.4.1",
-    "prettier": "~2.0.5",
+    "prettier": "~2.3.2",
     "react": "^16.13.1",
     "react-native": "^0.62.2",
     "replace-in-file": "^6.1.0",
@@ -56,7 +56,7 @@
     "ts-node": "^10.2.0",
     "tsc-watch": "^4.2.9",
     "typed-emitter": "^1.0.0",
-    "typescript": "~3.8.3",
+    "typescript": "~4.4.2",
     "wsrun": "^5.2.1"
   },
   "workspaces": {
