+++ packages/@magic-sdk/commons/package.json
@@ -21,7 +21,7 @@
   "peerDependencies": {
     "@magic-sdk/provider": ">=4.3.0",
     "@magic-sdk/types": ">=3.1.1",
-    "tslib": "^2.0.3"
+    "tslib": "^2.3.1"
   },
   "gitHead": "1ef062ea699d48d5e9a9375a93b7c147632b05ca"
 }
