+++ packages/@magic-sdk/provider/package.json
@@ -22,7 +22,7 @@
     "@magic-sdk/types": "^4.0.0",
     "eventemitter3": "^4.0.4",
     "semver": "^7.3.2",
-    "tslib": "^2.0.3",
+    "tslib": "^2.3.1",
     "web3-core": "1.5.2"
   },
   "peerDependencies": {
