+++ packages/@magic-sdk/provider/src/core/sdk.ts
@@ -92,7 +92,7 @@ function prepareExtensions(this: SDKBase, options?: MagicSDKAdditionalConfigurat
 
 export interface MagicSDKAdditionalConfiguration<
   TCustomExtName extends string = string,
-  TExt extends Extension<string>[] | { [P in TCustomExtName]: Extension<string> } = any
+  TExt extends Extension<string>[] | { [P in TCustomExtName]: Extension<string> } = any,
 > {
   endpoint?: string;
   locale?: SupportedLocale;
