+++ packages/@magic-sdk/react-native/package.json
@@ -32,7 +32,7 @@
     "localforage-driver-memory": "^1.0.5",
     "lodash": "^4.17.19",
     "process": "~0.11.10",
-    "tslib": "^2.0.3",
+    "tslib": "^2.3.1",
     "whatwg-url": "~8.1.0"
   },
   "devDependencies": {
