+++ packages/@magic-sdk/types/package.json
@@ -19,7 +19,7 @@
   "module": "dist/es/index.js",
   "types": "dist/es/index.d.ts",
   "dependencies": {
-    "tslib": "^2.0.3"
+    "tslib": "^2.3.1"
   },
   "gitHead": "1ef062ea699d48d5e9a9375a93b7c147632b05ca"
 }
