+++ packages/magic-sdk/package.json
@@ -33,7 +33,7 @@
     "localforage": "^1.7.4",
     "localforage-driver-memory": "^1.0.5",
     "regenerator-runtime": "0.13.5",
-    "tslib": "^2.0.3"
+    "tslib": "^2.3.1"
   },
   "gitHead": "1ef062ea699d48d5e9a9375a93b7c147632b05ca"
 }
