+++ scripts/bin/clean.ts
@@ -4,7 +4,7 @@ import execa from 'execa';
 import meow from 'meow';
 import ora from 'ora';
 import chalk from 'chalk';
-import { runAsyncProcess } from './utils/run-async-process';
+import { runAsyncProcess } from '../utils/run-async-process';
 
 const spinner = ora();
