+++ scripts/bin/inject-env.ts
@@ -6,7 +6,7 @@
 
 import { replaceInFile } from 'replace-in-file';
 import path from 'path';
-import { runAsyncProcess } from './utils/run-async-process';
+import { runAsyncProcess } from '../utils/run-async-process';
 
 /**
  * Environment variables to be interpolated into the built files. Interpolations
@@ -15,8 +15,8 @@ import { runAsyncProcess } from './utils/run-async-process';
  * (FYI: it's best to encapsulate the interpolation as a string, i.e.: `"%VARIABLE_NAME%"`).
  */
 const environment = {
-  WEB_VERSION: require(path.resolve(__dirname, '../packages/magic-sdk/package.json')).version,
-  REACT_NATIVE_VERSION: require(path.resolve(__dirname, '../packages/@magic-sdk/react-native/package.json')).version,
+  WEB_VERSION: require(path.resolve(__dirname, '../../packages/magic-sdk/package.json')).version,
+  REACT_NATIVE_VERSION: require(path.resolve(__dirname, '../../packages/@magic-sdk/react-native/package.json')).version,
 };
 
 async function main() {
